Project 5/requirements.txt
livekit-agents
livekit-plugins-deepgram
livekit-plugins-elevenlabs
livekit-plugins-openai
livekit-plugins-silero
python-dotenv
